kafka/server: add offset_store with the offset metadata types
Introduces offset_store.h and moves some things from group to there:

- various record version ints
- metadata structs for transactions
- metadata for offsets
- some using declarations for internal containers

Includes aliases for some of those that were consumed from outside group as
in `group::offset_metadata`.
